UI: moving ui to a separate file
just like with geojson polygon, I decide to move the branca legend and ui elements to a separate py file and import it so the main code file is cleaner and easier to read/wrte/navigate
